# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BudgetTracker API
  version: 1.0.0
  description: |
    BudgetTracker backend API secured by Firebase ID tokens.
    All protected endpoints require: Authorization: Bearer <FIREBASE_ID_TOKEN>
servers:
  - url: http://localhost:5000
    description: Local Development
  - url: https://budgettrackerapi-muxo.onrender.com
    description: Render Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID Token (NOT your own JWT)
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        code:
          type: string
      required: [error]
    User:
      type: object
      properties:
        _id:
          type: string
          example: "687848e1dd2446f74ce94a39"
        uid:
          type: string
          example: "irw4Vfd6DYWgG4ymLCpLHnMMxom2"
        email:
          type: string
          example: "user@example.com"
        displayName:
          type: string
          example: ""
        phoneNumber:
          type: string
          example: ""
        photoURL:
          type: string
          example: ""
        role:
          type: string
          example: "user"
        preferences:
          type: object
          properties:
            currency:
              type: string
              example: "CAD"
            theme:
              type: string
              example: "light"
        lastLogin:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register (idempotent)
      description: Creates Mongo user document if missing.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                photoURL:
                  type: string
                phoneNumber:
                  type: string
      responses:
        "200":
          description: User already registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "400":
          description: Missing fields
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Invalid Firebase token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Registration error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Invalid Firebase token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: User not found (register first)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "400":
          description: Login failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh server-side session info (updates lastLogin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Invalid Firebase token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: User not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Token refresh failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (revokes Firebase refresh tokens)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
        "401":
          description: Invalid Firebase token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Logout failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...